+++ python/pyngraph/ops/abs.cpp
@@ -14,9 +14,10 @@
 // limitations under the License.
 //*****************************************************************************
 
-#include "ngraph/op/abs.hpp"
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
+
+#include "ngraph/op/abs.hpp"
 #include "pyngraph/ops/abs.hpp"
 
 namespace py = pybind11;