+++ python/pyngraph/ops/concat.cpp
@@ -25,10 +25,8 @@ namespace py = pybind11;
 
 void regclass_pyngraph_op_Concat(py::module m)
 {
-    py::class_<ngraph::op::Concat,
-               std::shared_ptr<ngraph::op::Concat>,
-               ngraph::op::util::RequiresTensorViewArgs>
-        concat(m, "Concat");
+    py::class_<ngraph::op::Concat, std::shared_ptr<ngraph::op::Concat>, ngraph::op::Op> concat(
+        m, "Concat");
     concat.doc() = "ngraph.impl.op.Concat wraps ngraph::op::Concat";
     concat.def(py::init<const ngraph::NodeVector&, size_t>());
 }