+++ python/pyngraph/ops/convert.cpp
@@ -16,7 +16,7 @@
 
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
-//#include <string>
+
 #include "ngraph/op/convert.hpp"
 #include "pyngraph/ops/convert.hpp"
 
@@ -24,10 +24,8 @@ namespace py = pybind11;
 
 void regclass_pyngraph_op_Convert(py::module m)
 {
-    py::class_<ngraph::op::Convert,
-               std::shared_ptr<ngraph::op::Convert>,
-               ngraph::op::util::UnaryElementwise>
-        convert(m, "Convert");
+    py::class_<ngraph::op::Convert, std::shared_ptr<ngraph::op::Convert>, ngraph::op::Op> convert(
+        m, "Convert");
     convert.doc() = "ngraph.impl.op.Convert wraps ngraph::op::Convert";
     convert.def(py::init<const std::shared_ptr<ngraph::Node>&, const ngraph::element::Type&>());
 }