+++ python/pyngraph/shape.cpp
@@ -20,6 +20,7 @@
 
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
+
 #include "ngraph/shape.hpp" // ngraph::Shape
 #include "pyngraph/shape.hpp"