+++ python/test/ngraph/test_reduction.py
@@ -17,7 +17,7 @@ import numpy as np
 import pytest
 
 import ngraph as ng
-from test.ngraph.util import run_op_node
+from test.ngraph.util import run_op_node, get_runtime
 
 
 @pytest.mark.parametrize('ng_api_helper, numpy_function, reduction_axes', [
@@ -43,6 +43,56 @@ def test_reduction_ops(ng_api_helper, numpy_function, reduction_axes):
     expected = numpy_function(input_data, axis=reduction_axes)
     result = run_op_node([input_data], ng_api_helper, reduction_axes)
     assert np.allclose(result, expected)
+
+
+@pytest.config.gpu_skip(reason='Not implemented')
+def test_argmax():
+    runtime = get_runtime()
+    input_x = ng.constant(np.array([[9, 2, 10],
+                                    [12, 8, 4],
+                                    [6, 1, 5],
+                                    [3, 11, 7]], dtype=np.float32))
+    model = runtime.computation(ng.argmax(input_x, 0))
+    result = model()
+    assert np.allclose(result,
+                       np.array([1, 3, 0], dtype=np.int32))
+
+
+@pytest.config.gpu_skip(reason='Not implemented')
+def test_argmin():
+    runtime = get_runtime()
+    input_x = ng.constant(np.array([[12, 2, 10],
+                                    [9, 8, 4],
+                                    [6, 1, 5],
+                                    [3, 11, 7]], dtype=np.float32))
+    model = runtime.computation(ng.argmin(input_x, 0))
+    result = model()
+    assert np.allclose(result,
+                       np.array([3, 2, 1], dtype=np.int32))
+
+
+@pytest.config.gpu_skip(reason='Not implemented')
+def test_topk():
+    runtime = get_runtime()
+    input_x = ng.constant(np.array([[9, 2, 10],
+                                    [12, 8, 4],
+                                    [6, 1, 5],
+                                    [3, 11, 7]], dtype=np.float32))
+    comp_topk = ng.topk(input_x, 4, 0)
+    model0 = runtime.computation(ng.get_output_element(comp_topk, 0))
+    result0 = model0()
+    assert np.allclose(result0,
+                       np.array([[1, 3, 0],
+                                 [0, 1, 3],
+                                 [2, 0, 2],
+                                 [3, 2, 1]], dtype=np.int32))
+    model1 = runtime.computation(ng.get_output_element(comp_topk, 1))
+    result1 = model1()
+    assert np.allclose(result1,
+                       np.array([[12, 11, 10],
+                                 [9, 8, 7],
+                                 [6, 2, 5],
+                                 [3, 1, 4]], dtype=np.float32))
 
 
 def test_reduce():