+++ src/ngraph/frontend/onnx_import/op/sub.hpp
@@ -20,6 +20,7 @@
 #include "ngraph/op/subtract.hpp"
 
 #include "core/node.hpp"
+#include "utils/broadcasting.hpp"
 
 namespace ngraph
 {
@@ -29,7 +30,8 @@ namespace ngraph
         {
             inline NodeVector sub(const Node& node)
             {
-                NodeVector ng_inputs{node.get_ng_inputs()};
+                NodeVector ng_inputs{
+                    numpy_style_broadcast_for_binary_operation(node.get_ng_inputs())};
                 return {std::make_shared<ngraph::op::Subtract>(ng_inputs.at(0), ng_inputs.at(1))};
             }