+++ src/ngraph/runtime/cpu/pass/cpu_layout.cpp
@@ -46,12 +46,16 @@
 #include "ngraph/runtime/cpu/mkldnn_utils.hpp"
 #include "ngraph/runtime/cpu/op/batch_norm_relu.hpp"
 #include "ngraph/runtime/cpu/op/bounded_relu.hpp"
+#include "ngraph/runtime/cpu/op/conv_add.hpp"
 #include "ngraph/runtime/cpu/op/conv_bias.hpp"
 #include "ngraph/runtime/cpu/op/conv_relu.hpp"
 #include "ngraph/runtime/cpu/op/convert_layout.hpp"
+#include "ngraph/runtime/cpu/op/dequantize.hpp"
 #include "ngraph/runtime/cpu/op/group_conv.hpp"
 #include "ngraph/runtime/cpu/op/lstm.hpp"
 #include "ngraph/runtime/cpu/op/max_pool_with_indices.hpp"
+#include "ngraph/runtime/cpu/op/quantized_avg_pool.hpp"
+#include "ngraph/runtime/cpu/op/quantized_max_pool.hpp"
 #include "ngraph/runtime/cpu/op/rnn.hpp"
 
 using namespace std;
@@ -173,8 +177,8 @@ void runtime::cpu::pass::CPULayout::set_native_layouts(
     {
         const auto& output = input.get_output();
         auto tv = output.get_tensor_view();
-        auto et = tv->get_tensor_view_type()->get_element_type();
-        auto shape = tv->get_tensor_view_type()->get_shape();
+        auto et = tv->get_element_type();
+        auto shape = tv->get_shape();
         auto tvl = tv->get_tensor_view_layout();
         auto cpu_tvl = dynamic_cast<runtime::cpu::LayoutDescriptor*>(tvl.get());
 
@@ -242,8 +246,8 @@ void runtime::cpu::pass::CPULayout::set_native_layouts(
             continue;
         }
 
-        auto shape = tv->get_tensor_view_type()->get_shape();
-        auto et = tv->get_tensor_view_type()->get_element_type();
+        auto shape = tv->get_shape();
+        auto et = tv->get_element_type();
         auto layout = std::make_shared<ngraph::runtime::cpu::LayoutDescriptor>(*tv);
         if (mkldnn_utils::can_create_mkldnn_md(shape, layout->get_strides(), et))
         {
@@ -471,6 +475,26 @@ namespace ngraph
                     else
                     {
                         set_native_layouts(external_function, node);
+                    }
+                }
+
+                template <>
+                void CPULayout::LAYOUT_DECL(ngraph::op::ConvolutionAdd)
+                {
+                    if (mkldnn_utils::use_mkldnn_kernel(node.get()))
+                    {
+                        vector<memory::desc> i_mds;
+                        vector<memory::desc> o_mds;
+                        ConvolutionLayout<ngraph::op::ConvolutionAdd, false, false>(
+                            node, i_mds, o_mds);
+                        // Force second input to sum to use the same layout as convolution output
+                        i_mds.push_back(o_mds[0]);
+                        node = insert_input_conversions(external_function, node, i_mds);
+                        set_output_layouts(node, o_mds);
+                    }
+                    else
+                    {
+                        throw ngraph_error("ConvolutionAdd only supported in MKLDNN for now");
                     }
                 }
 
@@ -706,65 +730,71 @@ namespace ngraph
                     }
                 }
 
+                template <typename T>
+                void AvgPoolLayout(std::shared_ptr<ngraph::Node> node,
+                                   vector<memory::desc>& i_mds,
+                                   vector<memory::desc>& o_mds)
+                {
+                    auto avg_pool = static_cast<const ngraph::op::AvgPool*>(node.get());
+
+                    auto arg0_shape = node->get_input_shape(0);
+                    auto result_shape = node->get_output_shape(0);
+                    auto filter_shape = avg_pool->get_window_shape();
+                    auto filter_strides = avg_pool->get_window_movement_strides();
+                    auto padding_below = avg_pool->get_padding_below();
+                    auto padding_above = avg_pool->get_padding_above();
+
+                    memory::data_type et =
+                        mkldnn_utils::get_mkldnn_data_type(node->get_input_element_type(0));
+
+                    algorithm algorithm_enumerator =
+                        avg_pool->get_include_padding_in_avg_computation()
+                            ? algorithm::pooling_avg_include_padding
+                            : algorithm::pooling_avg_exclude_padding;
+
+                    memory::dims mkldnn_arg0_shape(arg0_shape.begin(), arg0_shape.end());
+                    memory::dims mkldnn_result_shape(result_shape.begin(), result_shape.end());
+                    memory::dims mkldnn_filter_shape(filter_shape.begin(), filter_shape.end());
+                    memory::dims mkldnn_filter_strides(filter_strides.begin(),
+                                                       filter_strides.end());
+                    memory::dims mkldnn_padding_below(padding_below.begin(), padding_below.end());
+                    memory::dims mkldnn_padding_above(padding_above.begin(), padding_above.end());
+
+                    auto input_desc = mkldnn_utils::get_input_mkldnn_md(node.get(), 0);
+                    auto result_desc = memory::desc(mkldnn_result_shape, et, memory::format::any);
+
+                    try
+                    {
+                        auto prim_desc =
+                            pooling_forward::primitive_desc({prop_kind::forward_inference,
+                                                             algorithm_enumerator,
+                                                             input_desc,
+                                                             result_desc,
+                                                             mkldnn_filter_strides,
+                                                             mkldnn_filter_shape,
+                                                             mkldnn_padding_below,
+                                                             mkldnn_padding_above,
+                                                             padding_kind::zero},
+                                                            mkldnn_utils::global_cpu_engine);
+                        i_mds.push_back(input_desc);
+                        o_mds.push_back(prim_desc.dst_primitive_desc().desc());
+                    }
+                    catch (const mkldnn::error& e)
+                    {
+                        throw ngraph_error("MKLDNN Unsupported pooling layout" +
+                                           to_string(input_desc.data.format) + e.message);
+                    }
+                }
+
                 template <>
                 void CPULayout::LAYOUT_DECL(ngraph::op::AvgPool)
                 {
                     if (mkldnn_utils::use_mkldnn_kernel(node.get()))
                     {
-                        auto avg_pool = static_cast<const ngraph::op::AvgPool*>(node.get());
-
-                        auto arg0_shape = node->get_input_shape(0);
-                        auto result_shape = node->get_output_shape(0);
-                        auto filter_shape = avg_pool->get_window_shape();
-                        auto filter_strides = avg_pool->get_window_movement_strides();
-                        auto padding_below = avg_pool->get_padding_below();
-                        auto padding_above = avg_pool->get_padding_above();
-
-                        memory::data_type et =
-                            mkldnn_utils::get_mkldnn_data_type(node->get_input_element_type(0));
-
-                        algorithm algorithm_enumerator =
-                            avg_pool->get_include_padding_in_avg_computation()
-                                ? algorithm::pooling_avg_include_padding
-                                : algorithm::pooling_avg_exclude_padding;
-
-                        memory::dims mkldnn_arg0_shape(arg0_shape.begin(), arg0_shape.end());
-                        memory::dims mkldnn_result_shape(result_shape.begin(), result_shape.end());
-                        memory::dims mkldnn_filter_shape(filter_shape.begin(), filter_shape.end());
-                        memory::dims mkldnn_filter_strides(filter_strides.begin(),
-                                                           filter_strides.end());
-                        memory::dims mkldnn_padding_below(padding_below.begin(),
-                                                          padding_below.end());
-                        memory::dims mkldnn_padding_above(padding_above.begin(),
-                                                          padding_above.end());
-
-                        auto input_desc = mkldnn_utils::get_input_mkldnn_md(node.get(), 0);
-                        auto result_desc =
-                            memory::desc(mkldnn_result_shape, et, memory::format::any);
-
                         vector<memory::desc> i_mds;
                         vector<memory::desc> o_mds;
-                        try
-                        {
-                            auto prim_desc =
-                                pooling_forward::primitive_desc({prop_kind::forward_inference,
-                                                                 algorithm_enumerator,
-                                                                 input_desc,
-                                                                 result_desc,
-                                                                 mkldnn_filter_strides,
-                                                                 mkldnn_filter_shape,
-                                                                 mkldnn_padding_below,
-                                                                 mkldnn_padding_above,
-                                                                 padding_kind::zero},
-                                                                mkldnn_utils::global_cpu_engine);
-                            i_mds.push_back(input_desc);
-                            o_mds.push_back(prim_desc.dst_primitive_desc().desc());
-                        }
-                        catch (const mkldnn::error& e)
-                        {
-                            throw ngraph_error("MKLDNN Unsupported pooling layout" +
-                                               to_string(input_desc.data.format) + e.message);
-                        }
+
+                        AvgPoolLayout<ngraph::op::AvgPool>(node, i_mds, o_mds);
 
                         node = insert_input_conversions(external_function, node, i_mds);
                         set_output_layouts(node, o_mds);
@@ -910,6 +940,73 @@ namespace ngraph
                     {
                         throw ngraph_error("MKLDNN Unsupported pooling fwd layout" +
                                            to_string(input_desc.data.format) + e.message);
+                    }
+                }
+
+                template <>
+                void CPULayout::LAYOUT_DECL(ngraph::op::QuantizedMaxPool)
+                {
+                    if (mkldnn_utils::use_mkldnn_kernel(node.get()))
+                    {
+                        vector<memory::desc> i_mds;
+                        vector<memory::desc> o_mds;
+
+                        MaxPoolLayout<ngraph::op::QuantizedMaxPool, prop_kind::forward_inference>(
+                            node, i_mds, o_mds);
+
+                        auto min_input_md = mkldnn_utils::create_default_mkldnn_md(
+                            node.get(), 1, false, memory::format::x);
+                        auto max_input_md = mkldnn_utils::create_default_mkldnn_md(
+                            node.get(), 2, false, memory::format::x);
+                        auto min_output_md = mkldnn_utils::create_default_mkldnn_md(
+                            node.get(), 1, true, memory::format::x);
+                        auto max_output_md = mkldnn_utils::create_default_mkldnn_md(
+                            node.get(), 2, true, memory::format::x);
+
+                        i_mds.push_back(min_input_md);
+                        i_mds.push_back(max_input_md);
+                        o_mds.push_back(min_output_md);
+                        o_mds.push_back(max_output_md);
+
+                        node = insert_input_conversions(external_function, node, i_mds);
+                        set_output_layouts(node, o_mds);
+                    }
+                    else
+                    {
+                        set_native_layouts(external_function, node);
+                    }
+                }
+
+                template <>
+                void CPULayout::LAYOUT_DECL(ngraph::op::QuantizedAvgPool)
+                {
+                    if (mkldnn_utils::use_mkldnn_kernel(node.get()))
+                    {
+                        vector<memory::desc> i_mds;
+                        vector<memory::desc> o_mds;
+
+                        AvgPoolLayout<ngraph::op::QuantizedAvgPool>(node, i_mds, o_mds);
+
+                        auto min_input_md = mkldnn_utils::create_default_mkldnn_md(
+                            node.get(), 1, false, memory::format::x);
+                        auto max_input_md = mkldnn_utils::create_default_mkldnn_md(
+                            node.get(), 2, false, memory::format::x);
+                        auto min_output_md = mkldnn_utils::create_default_mkldnn_md(
+                            node.get(), 1, true, memory::format::x);
+                        auto max_output_md = mkldnn_utils::create_default_mkldnn_md(
+                            node.get(), 2, true, memory::format::x);
+
+                        i_mds.push_back(min_input_md);
+                        i_mds.push_back(max_input_md);
+                        o_mds.push_back(min_output_md);
+                        o_mds.push_back(max_output_md);
+
+                        node = insert_input_conversions(external_function, node, i_mds);
+                        set_output_layouts(node, o_mds);
+                    }
+                    else
+                    {
+                        set_native_layouts(external_function, node);
                     }
                 }
 
@@ -1538,6 +1635,20 @@ namespace ngraph
                         set_native_layouts(external_function, node);
                     }
                 }
+
+                template <>
+                void CPULayout::LAYOUT_DECL(ngraph::op::Dequantize)
+                {
+                    if (mkldnn_utils::use_mkldnn_kernel(node.get()))
+                    {
+                        //TODO : Propogate Layout
+                        set_native_layouts(external_function, node);
+                    }
+                    else
+                    {
+                        throw ngraph_error("Dequantized op is only supported in MKLDNN for now.");
+                    }
+                }
             }
         }
     }
@@ -1559,6 +1670,10 @@ static const runtime::cpu::pass::LayoutOpMap s_dispatcher{
     {TI(ngraph::op::ConvolutionBackpropFilters),
      &runtime::cpu::pass::CPULayout::layout<ngraph::op::ConvolutionBackpropFilters>},
     {TI(ngraph::op::MaxPool), &runtime::cpu::pass::CPULayout::layout<ngraph::op::MaxPool>},
+    {TI(ngraph::op::QuantizedMaxPool),
+     &runtime::cpu::pass::CPULayout::layout<ngraph::op::QuantizedMaxPool>},
+    {TI(ngraph::op::QuantizedAvgPool),
+     &runtime::cpu::pass::CPULayout::layout<ngraph::op::QuantizedAvgPool>},
     {TI(ngraph::op::MaxPoolWithIndices),
      &runtime::cpu::pass::CPULayout::layout<ngraph::op::MaxPoolWithIndices>},
     {TI(ngraph::op::MaxPoolBackprop),
@@ -1593,6 +1708,9 @@ static const runtime::cpu::pass::LayoutOpMap s_dispatcher{
     {TI(ngraph::op::Rnn), &runtime::cpu::pass::CPULayout::layout<ngraph::op::Rnn>},
     {TI(ngraph::op::Softmax), &runtime::cpu::pass::CPULayout::layout<ngraph::op::Softmax>},
     {TI(ngraph::op::BoundedRelu), &runtime::cpu::pass::CPULayout::layout<ngraph::op::BoundedRelu>},
+    {TI(ngraph::op::ConvolutionAdd),
+     &runtime::cpu::pass::CPULayout::layout<ngraph::op::ConvolutionAdd>},
+    {TI(ngraph::op::Dequantize), &runtime::cpu::pass::CPULayout::layout<ngraph::op::Dequantize>},
 };
 
 bool runtime::cpu::pass::CPULayout::run_on_call_graph(const std::list<std::shared_ptr<Node>>& nodes)